# src/03-mcp-servers/openapi/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "EcommerceAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/customers/id/{customer_id}": {
            "get": {
                "summary": "Get Customer By Id",
                "description": "Get customer by ID",
                "operationId": "get_customer_by_id",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Customer Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                }
            }
        },
        "/customers/name/{customer_name}": {
            "get": {
                "summary": "Get Customer By Name",
                "description": "Get customer by name",
                "operationId": "get_customer_by_name",
                "parameters": [
                    {
                        "name": "customer_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Customer Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                }
            }
        },
        "/products/all": {
            "get": {
                "summary": "Get All Products",
                "description": "Get all products",
                "operationId": "get_all_products",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    },
                                    "type": "array",
                                    "title": "Response Get All Products"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/discounts/all": {
            "get": {
                "summary": "Get All Discounts",
                "description": "Get all discounts",
                "operationId": "get_all_discounts",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Discount"
                                    },
                                    "type": "array",
                                    "title": "Response Get All Discounts"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/orders/id/{order_id}": {
            "get": {
                "summary": "Get Order By Id",
                "description": "Get order by ID",
                "operationId": "get_order_by_id",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Order Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                }
            }
        },
        "/order/update": {
            "post": {
                "summary": "Update Order",
                "description": "Update existing order",
                "operationId": "update_order",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "boolean",
                                    "title": "Response Update Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/get_closest_inventory_location/{customer_name}": {
            "get": {
                "summary": "Get Closest Inventory Location",
                "description": "Get closest inventory location based on customer name",
                "operationId": "get_closest_inventory_location",
                "parameters": [
                    {
                        "name": "customer_name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Customer Name"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "title": "Response Get Closest Inventory Location"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                }
            }
        },
        "/inventory/{product_id}": {
            "get": {
                "summary": "Get Inventory By Product Id",
                "description": "Get available inventory by product ID",
                "operationId": "get_inventory_by_product_id",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Product Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductInventory"
                                    },
                                    "title": "Response Get Inventory By Product Id"
                                }
                            }
                        }
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        },
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Customer": {
                "properties": {
                    "customer_id": {
                        "type": "string",
                        "title": "Customer Id"
                    },
                    "customer_name": {
                        "type": "string",
                        "title": "Customer Name"
                    },
                    "customer_address": {
                        "type": "string",
                        "title": "Customer Address"
                    },
                    "customer_phone": {
                        "type": "string",
                        "title": "Customer Phone"
                    },
                    "customer_email": {
                        "type": "string",
                        "title": "Customer Email"
                    },
                    "customer_discount": {
                        "items": {
                            "$ref": "#/components/schemas/CustomerDiscount"
                        },
                        "type": "array",
                        "title": "Customer Discount"
                    }
                },
                "type": "object",
                "required": [
                    "customer_id",
                    "customer_name",
                    "customer_address",
                    "customer_phone",
                    "customer_email"
                ],
                "title": "Customer"
            },
            "CustomerDiscount": {
                "properties": {
                    "customer_id": {
                        "type": "string",
                        "title": "Customer Id"
                    },
                    "discount_id": {
                        "type": "string",
                        "title": "Discount Id"
                    },
                    "discount_name": {
                        "type": "string",
                        "title": "Discount Name"
                    },
                    "discount_price": {
                        "type": "number",
                        "title": "Discount Price"
                    },
                    "product_id": {
                        "type": "string",
                        "title": "Product Id"
                    },
                    "discount_volume": {
                        "type": "integer",
                        "title": "Discount Volume"
                    }
                },
                "type": "object",
                "required": [
                    "customer_id",
                    "discount_id",
                    "discount_name",
                    "discount_price",
                    "product_id",
                    "discount_volume"
                ],
                "title": "CustomerDiscount"
            },
            "Discount": {
                "properties": {
                    "discount_id": {
                        "type": "string",
                        "title": "Discount Id"
                    },
                    "discount_name": {
                        "type": "string",
                        "title": "Discount Name"
                    },
                    "discount_price": {
                        "type": "number",
                        "title": "Discount Price"
                    },
                    "product_id": {
                        "type": "string",
                        "title": "Product Id"
                    },
                    "discount_volume": {
                        "type": "integer",
                        "title": "Discount Volume"
                    }
                },
                "type": "object",
                "required": [
                    "discount_id",
                    "discount_name",
                    "discount_price",
                    "product_id",
                    "discount_volume"
                ],
                "title": "Discount"
            },
            "Message": {
                "properties": {
                    "message": {
                        "type": "string",
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "message"
                ],
                "title": "Message"
            },
            "Order": {
                "properties": {
                    "customer_id": {
                        "type": "string",
                        "title": "Customer Id"
                    },
                    "order_id": {
                        "type": "string",
                        "title": "Order Id"
                    },
                    "order_date": {
                        "type": "string",
                        "title": "Order Date"
                    },
                    "order_status": {
                        "type": "string",
                        "title": "Order Status"
                    },
                    "fill_date": {
                        "type": "string",
                        "title": "Fill Date"
                    },
                    "fill_strategy": {
                        "type": "string",
                        "title": "Fill Strategy"
                    },
                    "order_items": {
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        },
                        "type": "array",
                        "title": "Order Items"
                    }
                },
                "type": "object",
                "required": [
                    "customer_id",
                    "order_id",
                    "order_date",
                    "order_status",
                    "fill_date"
                ],
                "title": "Order"
            },
            "Product": {
                "properties": {
                    "product_id": {
                        "type": "string",
                        "title": "Product Id"
                    },
                    "product_name": {
                        "type": "string",
                        "title": "Product Name"
                    },
                    "list_price": {
                        "type": "number",
                        "title": "List Price"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "features": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Features"
                    }
                },
                "type": "object",
                "required": [
                    "product_id",
                    "product_name",
                    "list_price",
                    "description",
                    "features"
                ],
                "title": "Product"
            },
            "ProductInventory": {
                "properties": {
                    "product_id": {
                        "type": "string",
                        "title": "Product Id"
                    },
                    "product_name": {
                        "type": "string",
                        "title": "Product Name"
                    },
                    "volume": {
                        "type": "integer",
                        "title": "Volume"
                    },
                    "location": {
                        "type": "string",
                        "title": "Location"
                    }
                },
                "type": "object",
                "required": [
                    "product_id",
                    "product_name",
                    "volume",
                    "location"
                ],
                "title": "ProductInventory"
            }
        }
    },
    "servers": [
        {
            "url": "https://l66ctpb6-8000.euw.devtunnels.ms"
        }
    ]
}